+++ examples/server_instantiation.cpp
@@ -16,9 +16,9 @@ int main() {
     nodeMamalType.writeDescription({"en-US", "A mamal"});
 
     auto nodeMamalTypeAge = nodeMamalType.addVariable({1, 10001}, "Age");
+    nodeMamalTypeAge.addModellingRule(opcua::ModellingRule::Mandatory);  // create on new instance
     nodeMamalTypeAge.writeDisplayName({"en-US", "Age"});
     nodeMamalTypeAge.writeDescription({"en-US", "This mamals age in months"});
-    nodeMamalTypeAge.writeModellingRule(opcua::ModellingRule::Mandatory);  // create on new instance
     nodeMamalTypeAge.writeScalar<uint32_t>(0);  // default age
 
     auto nodeDogType = nodeMamalType.addObjectType({1, 10002}, "DogType");
@@ -26,9 +26,9 @@ int main() {
     nodeDogType.writeDescription({"en-US", "A dog, subtype of mamal"});
 
     auto nodeDogTypeName = nodeDogType.addVariable({1, 10003}, "Name");
+    nodeDogTypeName.addModellingRule(opcua::ModellingRule::Mandatory);  // create on new instance
     nodeDogTypeName.writeDisplayName({"en-US", "Name"});
     nodeDogTypeName.writeDescription({"en-US", "This dogs name"});
-    nodeDogTypeName.writeModellingRule(opcua::ModellingRule::Mandatory);  // create on new instance
     nodeDogTypeName.writeScalar(opcua::String("unnamed dog"));  // default name
 
     /* Instatiate a dog named Bello: