+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Return reference from `ExpandedNodeId::getNodeId`
 - Pass `LocalizedText` instead of members (`locale`, `text`) to `Node::writeDisplayName`, `Node::writeDescription`, `services::writeDisplayName`, `services::writeDescription` (#29)
 - Remove `Server::getConfig` method
+- Rename `Node::writeModellingRule` -> `Node::addModellingRule`
 
 ### Fixed